refactor: extract deployTestTokenEcosystem to dedicated module
